# packages/api-schema/open-api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/settings": {
      "get": {
        "operationId": "SettingsController_getSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "object"
                  ],
                  "properties": {
                    "lockTimeoutMins": {
                      "type": [
                        "number"
                      ]
                    },
                    "departuresSiteId": {
                      "type": [
                        "number",
                        "null"
                      ]
                    },
                    "departuresSiteName": {
                      "type": [
                        "string"
                      ]
                    },
                    "departuresCount": {
                      "type": [
                        "number"
                      ]
                    },
                    "icaListId": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "lockTimeoutMins",
                    "departuresSiteId",
                    "departuresSiteName",
                    "departuresCount",
                    "icaListId"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": [
                        "object"
                      ],
                      "properties": {
                        "code": {
                          "type": [
                            "string"
                          ],
                          "enum": [
                            "INVALID_BODY"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    },
                    {
                      "type": [
                        "object"
                      ],
                      "properties": {
                        "code": {
                          "type": [
                            "string"
                          ],
                          "enum": [
                            "INVALID_PATH_PARAMETER"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    },
                    {
                      "type": [
                        "object"
                      ],
                      "properties": {
                        "code": {
                          "type": [
                            "string"
                          ],
                          "enum": [
                            "INVALID_QUERY_PARAMETER"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Settings"
        ]
      },
      "patch": {
        "operationId": "SettingsController_updateSettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": [
                  "object"
                ],
                "properties": {
                  "lockTimeoutMins": {
                    "type": [
                      "number"
                    ]
                  },
                  "departuresSiteId": {
                    "type": [
                      "number",
                      "null"
                    ]
                  },
                  "departuresSiteName": {
                    "type": [
                      "string"
                    ]
                  },
                  "departuresCount": {
                    "type": [
                      "number"
                    ]
                  },
                  "icaListId": {
                    "type": [
                      "string",
                      "null"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "object"
                  ],
                  "properties": {
                    "lockTimeoutMins": {
                      "type": [
                        "number"
                      ]
                    },
                    "departuresSiteId": {
                      "type": [
                        "number",
                        "null"
                      ]
                    },
                    "departuresSiteName": {
                      "type": [
                        "string"
                      ]
                    },
                    "departuresCount": {
                      "type": [
                        "number"
                      ]
                    },
                    "icaListId": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "lockTimeoutMins",
                    "departuresSiteId",
                    "departuresSiteName",
                    "departuresCount",
                    "icaListId"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": [
                        "object"
                      ],
                      "properties": {
                        "code": {
                          "type": [
                            "string"
                          ],
                          "enum": [
                            "INVALID_BODY"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    },
                    {
                      "type": [
                        "object"
                      ],
                      "properties": {
                        "code": {
                          "type": [
                            "string"
                          ],
                          "enum": [
                            "INVALID_PATH_PARAMETER"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    },
                    {
                      "type": [
                        "object"
                      ],
                      "properties": {
                        "code": {
                          "type": [
                            "string"
                          ],
                          "enum": [
                            "INVALID_QUERY_PARAMETER"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Settings"
        ]
      }
    },
    "/api/pages": {
      "get": {
        "operationId": "SettingsController_getPages",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array"
                  ],
                  "items": {
                    "type": [
                      "object"
                    ],
                    "properties": {
                      "id": {
                        "type": [
                          "string"
                        ]
                      },
                      "name": {
                        "type": [
                          "string"
                        ]
                      },
                      "position": {
                        "type": [
                          "number"
                        ]
                      },
                      "layout": {
                        "type": [
                          "array"
                        ],
                        "items": {
                          "type": [
                            "object"
                          ],
                          "properties": {
                            "id": {
                              "type": [
                                "string"
                              ]
                            },
                            "colStart": {
                              "type": [
                                "number"
                              ]
                            },
                            "rowStart": {
                              "type": [
                                "number"
                              ]
                            },
                            "colSpan": {
                              "type": [
                                "number"
                              ]
                            },
                            "rowSpan": {
                              "type": [
                                "number"
                              ]
                            }
                          },
                          "required": [
                            "id",
                            "colStart",
                            "rowStart",
                            "colSpan",
                            "rowSpan"
                          ],
                          "additionalProperties": true
                        }
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "position",
                      "layout"
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": [
                        "object"
                      ],
                      "properties": {
                        "code": {
                          "type": [
                            "string"
                          ],
                          "enum": [
                            "INVALID_BODY"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    },
                    {
                      "type": [
                        "object"
                      ],
                      "properties": {
                        "code": {
                          "type": [
                            "string"
                          ],
                          "enum": [
                            "INVALID_PATH_PARAMETER"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    },
                    {
                      "type": [
                        "object"
                      ],
                      "properties": {
                        "code": {
                          "type": [
                            "string"
                          ],
                          "enum": [
                            "INVALID_QUERY_PARAMETER"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Settings"
        ]
      },
      "put": {
        "operationId": "SettingsController_setPages",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": [
                  "array"
                ],
                "items": {
                  "type": [
                    "object"
                  ],
                  "properties": {
                    "id": {
                      "type": [
                        "string"
                      ]
                    },
                    "name": {
                      "type": [
                        "string"
                      ]
                    },
                    "position": {
                      "type": [
                        "number"
                      ]
                    },
                    "layout": {
                      "type": [
                        "array"
                      ],
                      "items": {
                        "type": [
                          "object"
                        ],
                        "properties": {
                          "id": {
                            "type": [
                              "string"
                            ]
                          },
                          "colStart": {
                            "type": [
                              "number"
                            ]
                          },
                          "rowStart": {
                            "type": [
                              "number"
                            ]
                          },
                          "colSpan": {
                            "type": [
                              "number"
                            ]
                          },
                          "rowSpan": {
                            "type": [
                              "number"
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "colStart",
                          "rowStart",
                          "colSpan",
                          "rowSpan"
                        ],
                        "additionalProperties": true
                      }
                    }
                  },
                  "required": [
                    "name",
                    "position",
                    "layout"
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": [
                    "array"
                  ],
                  "items": {
                    "type": [
                      "object"
                    ],
                    "properties": {
                      "id": {
                        "type": [
                          "string"
                        ]
                      },
                      "name": {
                        "type": [
                          "string"
                        ]
                      },
                      "position": {
                        "type": [
                          "number"
                        ]
                      },
                      "layout": {
                        "type": [
                          "array"
                        ],
                        "items": {
                          "type": [
                            "object"
                          ],
                          "properties": {
                            "id": {
                              "type": [
                                "string"
                              ]
                            },
                            "colStart": {
                              "type": [
                                "number"
                              ]
                            },
                            "rowStart": {
                              "type": [
                                "number"
                              ]
                            },
                            "colSpan": {
                              "type": [
                                "number"
                              ]
                            },
                            "rowSpan": {
                              "type": [
                                "number"
                              ]
                            }
                          },
                          "required": [
                            "id",
                            "colStart",
                            "rowStart",
                            "colSpan",
                            "rowSpan"
                          ],
                          "additionalProperties": true
                        }
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "position",
                      "layout"
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": [
                        "object"
                      ],
                      "properties": {
                        "code": {
                          "type": [
                            "string"
                          ],
                          "enum": [
                            "INVALID_BODY"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    },
                    {
                      "type": [
                        "object"
                      ],
                      "properties": {
                        "code": {
                          "type": [
                            "string"
                          ],
                          "enum": [
                            "INVALID_PATH_PARAMETER"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    },
                    {
                      "type": [
                        "object"
                      ],
                      "properties": {
                        "code": {
                          "type": [
                            "string"
                          ],
                          "enum": [
                            "INVALID_QUERY_PARAMETER"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Settings"
        ]
      }
    },
...
  },
  "info": {
    "title": "SmartFridge API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
...
}